Create Eventing Manager * Separate using EC reconciler logic into Eventing Manager. This makes the cod more unit testable
